+++ packages/terra-brand-footer/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* 'links' prop has been deprecated, will resolve to 'sections'
 
 2.24.0 - (May 26, 2020)
 ------------------